working on update

+++ src/controllers/ressortController.js
@@ -1,38 +1,49 @@
-const token2 = localStorage.getItem("token");
+var globalId;
+const tokenRessort = localStorage.getItem("token");
 
-const myHeaders = new Headers();
-myHeaders.append("Authorization", "Bearer " + token2);
+function getRowId(td) {
+    selectedRow = td.parentElement.parentElement;
+    globalId = selectedRow.cells[0].innerHTML;
+    return globalId;
+}
 
-fetch('http://127.0.0.1:3000/api/ressort', {
-    method: 'GET',
-    headers: myHeaders,
-    mode: 'cors',
-    cache: 'default'
-})
-    .then(res => res.json())
-    .then(data => {
-        console.log(data);
-        console.log(data.data.length);
+function token() {
+    var token = localStorage.getItem('token');
+    console.log(token);
+}
 
-        if (data.data.length > 0) {
-            var body = "";
+function getRessort() {
+    const myHeaders = new Headers();
+    myHeaders.append("Authorization", "Bearer " + tokenRessort);
 
-            data.data.forEach(i => {
-                body += "<tr>";
-                body += "<td>" + i.ressort_id + "</td>";
-                body += "<td>" + i.ressortnaam + "</td>";
-                body += "<td>" + i.district + "</td>";
-                body += `<td>
-                    <a class='modal-trigger' href='#modal_updateressort' title='Wijzigen' data-toggle='tooltip' style='cursor: pointer;' onclick='return getData(this)'><i class='small material-icons' style='color: #ffd600;'>edit</i></a>
-                    <a title='Verwijderen' data-toggle='tooltip' style='cursor: pointer;' onclick='return deleteCheck(this)'><i class='small material-icons' style='color: #c62828;'>delete</i></a>
-                </td>`;
-                body += "</tr>";
-            })
-
-            document.getElementById('rTableBody').innerHTML = body;
-        }
+    fetch('http://127.0.0.1:3000/api/ressort', {
+        method: 'GET',
+        headers: myHeaders,
+        mode: 'cors',
+        cache: 'default'
     })
-    .catch((err) => console.log(err))
+        .then(res => res.json())
+        .then(data => {
+            console.log(data);
+            console.log(data.data.length);
+            if (data.data.length > 0) {
+                var body = "";
+                data.data.forEach(i => {
+                    body += "<tr>";
+                    body += "<td>" + i.ressort_id + "</td>";
+                    body += "<td>" + i.ressortnaam + "</td>";
+                    body += "<td>" + i.district + "</td>";
+                    body += `<td>
+                    <a class='modal-trigger' href='#modal_updateressort' title='Wijzigen' data-toggle='tooltip' style='cursor: pointer;' onclick='return getData(this)'><i class='small material-icons' style='color: #ffd600;'>edit</i></a>
+                    <a title='Verwijderen' data-toggle='tooltip' style='cursor: pointer;' onclick='return deleteRessort(this)'><i class='small material-icons' style='color: #c62828;'>delete</i></a>
+                </td>`;
+                    body += "</tr>";
+                })
+                document.getElementById('rTableBody').innerHTML = body;
+            }
+        })
+        .catch((err) => console.log(err))
+}
 
 function createRessort() {
     let form = document.forms["ressortForm"];
@@ -46,7 +57,7 @@ function createRessort() {
     console.log(VALUE);
 
     const myHeaders = new Headers();
-    myHeaders.append("Authorization", "Bearer " + token2);
+    myHeaders.append("Authorization", "Bearer " + tokenRessort);
     myHeaders.append('Content-Type', 'application/json');
     myHeaders.append('Accept', 'application/json');
 
@@ -70,18 +81,13 @@ function createRessort() {
     return false;
 }
 
-function token() {
-    var token = localStorage.getItem('token');
-    console.log(token);
-}
+function deleteRessort(td) {
+    id = getRowId(td);
 
-function deleteCheck(td) {
-    selectedRow = td.parentElement.parentElement;
-    id = selectedRow.cells[0].innerHTML;
+    if (confirm('Bent u zeker?')) {
 
-    function deleteRessort() {
         const myHeaders = new Headers();
-        myHeaders.append('Authorization', 'Bearer ' + token2);
+        myHeaders.append('Authorization', 'Bearer ' + tokenRessort);
 
         fetch('http://127.0.0.1:3000/api/ressort/' + id, {
             method: 'DELETE',
@@ -98,14 +104,19 @@ function deleteCheck(td) {
 
         return false;
     }
-
-    if (confirm('Bent u zeker?')) {
-        deleteRessort();
-    }
 }
 
-function getData(td) {
+function getRessortByID(td) {
     selectedRow = td.parentElement.parentElement;
     document.getElementById('updateRessortnaam').value = selectedRow.cells[1].innerHTML;
-    // document.getElementById('updateDistrict').selected = selectedRow.cells[2].innerHTML;
+    document.getElementById('updateDistrict').selected = selectedRow.cells[2].innerHTML;
+
+    id = getRowId(td);
+    alert(id);
+    return id;
+}
+
+function updateRessort(td) {
+    id = getRowId(td);
+    alert(id);
 }
